queue-writer/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Redis support (use compatible version)
redis[hiredis]==4.6.0

# Configuration and serialization
PyYAML==6.0.1

# HTTP client (for health checks)
httpx==0.25.2

# Logging and monitoring
structlog==23.2.0

# Development and testing (optional, remove in production)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1
flake8==6.1.0

# Production server
gunicorn==21.2.0